+++ src/main/java/model/bridge/RandomBridgesGenerator.java
@@ -1,4 +1,4 @@
-package model;
+package model.bridge;
 
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.ArrayList;
