+++ src/test/java/model/ladder/LadderHeightTest.java
@@ -1,4 +1,4 @@
-package model;
+package model.ladder;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
