+++ src/test/java/model/ladder/LadderTest.java
@@ -1,8 +1,10 @@
-package model;
+package model.ladder;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
+import model.bridge.Bridge;
+import model.player.Players;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
