+++ src/test/java/model/player/PlayersTest.java
@@ -1,4 +1,4 @@
-package model;
+package model.player;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
